+++ composeApp/src/commonMain/kotlin/domain/RepositoryImpl.kt
@@ -7,7 +7,8 @@ private var listBook : List<Book_table> = listOf()
 private var bookImpl : Book_table? = null
 private var chapterImpl : Long = 1
 private var database : BibliaDatabase? = null
-private var verseImpl : Long = 0
+private var verseImpl : Int = 0
+
 class RepositoryImpl : BibliaDBRepository {
     override fun getBooks(): List<Book_table> {
         return listBook
@@ -41,11 +42,11 @@ class RepositoryImpl : BibliaDBRepository {
         chapterImpl = chapter
     }
 
-    override fun getVerse(): Long {
+    override fun getVerse(): Int {
         return verseImpl
     }
 
-    override fun setVerse(verse: Long) {
+    override fun setVerse(verse: Int) {
         verseImpl = verse
     }
 }
